Moved dollar one level up
